ml-service/requirements.txt
grpcio
grpcio-tools
protobuf

torch
torchvision

Pillow

numpy
setuptools
